rename platformio_override.ini
this file is supposed to be changed by users, but users' changes shall
never be included in the upstream project. also, it is cumbersome that
git marks the file as changed and that it has to be cleaned up when
switching branched etc.

rename the file to platformio_override_sample.ini. this way users who
actually want to use the platformio_override.ini have something to start
from. they only need to copy the sample file. using '_sample' within the
file name rather than manipulating the file ending is good for windows
users who might not be displaying file extensions in their file
explorer.