Implements safe dynamic limits
When an inverter has the safe limit timeout enabled, a new /api/limit/safe WebApi
is enabled that allow to setup dynamic non-persistent limits.

Those dynamic limits are reverted if they are not confirmed periodically before the
safe limit timeout expires.

This way, if the controller or the WiFi router dies, the limit is safely restored
to the default limit.