compress web app using brotli and serve as such
this saves ~38k of flash memory, since brotli is much more effective at
compressing the web application. as we serve the web application binary
as is, the browser needs to take care of decompression. so we have to
look at browser support: as it turns out, brotli compression is
available on all browsers since more than four years. as per
caniuse.com/brotli, the support for brotli is available for 97% of users
today. we can safely assume that all OpenDTU users use a browser with
support for brotli encoded content.

the gzip-compressed transpiled files in webapp_dist are removed from the
repo and the new brotli-compressed transpiled files are added to the
gitignore list, as build artifacts shall not be tracked.