Fix: disable temperature readings for ESP32-S3
reading the temperature is flawed and leads to unexpected reboots, which
we observed more and more often recently. since the user experience is
trash if the module resets unexpectedly from time to time, it seems to
be more important to prevent reboots where ever possible rather than
keeping the temperature readings.

closes #1708.