fix: bump ESPAsyncWebServer to v3.7.6
after OpenDTU v24.9.30 was released there was a subtle change in the ESPAsyncWebServer
which changed the internal data structure for AsyncResponseStream from cbuf to StringStream.
This introduced WDT resets when scraping at least the prometheus API endpoint [1].
ESPAsyncWebServer v3.7.6 addresses this issue by reverting the data structure back to cbuf.

[1] https://github.com/tbnobody/OpenDTU/issues/2535

Signed-off-by: Peter Lieven <pl@dlhnet.de>